+++ tests/unit_tests (deprecated)/test_riglib_source.py
@@ -8,7 +8,7 @@ import time
 
 class MockDataSourceSystem(source.DataSourceSystem):
     update_freq = 2000
-    dtype = np.dtype([("value", np.float)])
+    dtype = np.dtype([("value", float)])
     attr_to_test_access = 43
     
     state = 0
